Flow message is displaying in pairing screen
